+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 
+## v0.3.0 | 17 Jul 2022
+- Data can now be uploaded in UTC (=GMT) time with timezone info.
+- Data can now be downloaded in desired timezone, e.g. `GMT+1` for `CET` (uses pytz.timezones). 
+
+
 ## v0.2.0 | 28 Jun 2022
 - Renamed package to `dbc-influxdb` (previous name was `dbc`)